+++ packages/desktop-app/src/app/components/command-bar/command-bar.component.spec.ts
@@ -19,6 +19,13 @@ describe("CommandBarComponent", () => {
     const spyRepositoryService = jasmine.createSpyObj("Repository", {
       getProfiles: [],
       getColorTheme: () => constants.darkTheme,
+      setNotifications: () => {},
+      getNotifications: () => [],
+    });
+    const spyNotificationsService = jasmine.createSpyObj("NotificationService", {
+      setNotifications: () => {},
+      getNotifications: () => [],
+      getNotificationByUuid: () => {},
     });
     const spyLeappCoreService = jasmine.createSpyObj("LeappCoreService", [], {
       behaviouralSubjectService: spyBehaviouralSubjectService,
@@ -26,6 +33,7 @@ describe("CommandBarComponent", () => {
       awsCoreService: { getRegions: () => [] },
       namedProfileService: { getNamedProfiles: () => [] },
       teamService: { signedInUserState: { subscribe: () => {} }, workspaceState: { subscribe: () => {} } },
+      notificationService: spyNotificationsService,
     });
 
     const optionsService = { colorTheme: "dark-theme" };
@@ -42,7 +50,6 @@ describe("CommandBarComponent", () => {
 
     fixture = TestBed.createComponent(CommandBarComponent);
     component = fixture.componentInstance;
-    fixture.detectChanges();
 
     (component as any).subscription0 = {
       unsubscribe: () => {},
@@ -72,6 +79,12 @@ describe("CommandBarComponent", () => {
       unsubscribe: () => {},
     };
     (component as any).optionsService = { colorTheme: "dark-theme", workspaceService: { getWorkspace: () => new Workspace() } };
+
+    (component as any).notificationService.getNotificationByUuid = () => {};
+    const spy = spyOnProperty(component as any, "notifications", "get");
+    spy.and.returnValue([]);
+
+    fixture.detectChanges();
   });
 
   it("should create", () => {