+++ packages/desktop-app/src/app/components/components.module.ts
@@ -53,6 +53,7 @@ import { MatSelectModule } from "@angular/material/select";
 import { MatInputModule } from "@angular/material/input";
 import { MatProgressSpinnerModule } from "@angular/material/progress-spinner";
 import { ManageTeamWorkspacesDialogComponent } from "./dialogs/manage-team-workspaces-dialog/manage-team-workspaces-dialog.component";
+import { InfoDialogComponent } from "./dialogs/info-dialog/info-dialog.component";
 
 @NgModule({
   declarations: [
@@ -82,6 +83,7 @@ import { ManageTeamWorkspacesDialogComponent } from "./dialogs/manage-team-works
     SsmModalDialogComponent,
     ContextualMenuComponent,
     BottomBarComponent,
+    InfoDialogComponent,
     LoginTeamDialogComponent,
     ManageTeamWorkspacesDialogComponent,
   ],