+++ packages/desktop-app/src/app/components/dialogs/info-dialog/info-dialog.component.spec.ts
@@ -1,0 +1,28 @@
+import { ComponentFixture, TestBed } from "@angular/core/testing";
+
+import { InfoDialogComponent } from "./info-dialog.component";
+import { mustInjected } from "../../../../base-injectables";
+import { RouterTestingModule } from "@angular/router/testing";
+
+describe("InfoDialogComponent", () => {
+  let component: InfoDialogComponent;
+  let fixture: ComponentFixture<InfoDialogComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [InfoDialogComponent],
+      providers: [].concat(mustInjected()),
+      imports: [RouterTestingModule],
+    }).compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(InfoDialogComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it("should create", () => {
+    expect(component).toBeTruthy();
+  });
+});