+++ packages/desktop-app/src/app/components/tray-menu/tray-menu.component.ts
@@ -13,11 +13,10 @@ import { AwsIamRoleChainedSession } from "@noovolari/leapp-core/models/aws/aws-i
 import { WindowService } from "../../services/window.service";
 import { AwsCoreService } from "@noovolari/leapp-core/services/aws-core-service";
 import { AppNativeService } from "../../services/app-native.service";
-import { LeappBaseError } from "@noovolari/leapp-core/errors/leapp-base-error";
-import { MessageToasterService, ToastLevel } from "../../services/message-toaster.service";
-import { LoggedEntry, LoggedException, LogLevel, LogService } from "@noovolari/leapp-core/services/log-service";
+import { MessageToasterService } from "../../services/message-toaster.service";
+import { LoggedEntry, LogLevel, LogService } from "@noovolari/leapp-core/services/log-service";
 import { OperatingSystem } from "@noovolari/leapp-core/models/operating-system";
-import { LeappLinkError } from "@noovolari/leapp-core/errors/leapp-link-error";
+import { constants } from "@noovolari/leapp-core/models/constants";
 
 @Component({
   selector: "app-tray-menu",
@@ -33,10 +32,6 @@ export class TrayMenuComponent implements OnInit, OnDestroy {
   private loggingService: LogService;
   private sessionServiceFactory: SessionFactory;
   private behaviouralSubjectService: BehaviouralSubjectService;
-
-  private awsCliVersion: string;
-  private awsSsmPluginVersion: string;
-  private issueBody: string;
 
   private voices = [];
   private sessions = [];
@@ -60,7 +55,6 @@ export class TrayMenuComponent implements OnInit, OnDestroy {
     this.subscribed = this.behaviouralSubjectService.sessions$.subscribe(() => {
       this.generateMenu();
     });
-    this.getMetadata();
     await this.generateMenu();
   }
 
@@ -115,15 +109,17 @@ export class TrayMenuComponent implements OnInit, OnDestroy {
         label: "Join Slack Community",
         type: "normal",
         click: () => {
-          this.windowService.openExternalUrl("https://join.slack.com/t/noovolari/shared_invite/zt-opn8q98k-HDZfpJ2_2U3RdTnN~u_B~Q");
+          this.windowService.openExternalUrl(constants.slackUrl);
         },
       },
       {
         label: "Open Issue",
         type: "normal",
-        enabled: this.awsSsmPluginVersion && this.awsCliVersion && this.issueBody,
+        enabled: this.appService.awsSsmPluginVersion && this.appService.awsCliVersion && this.appService.issueBody,
         click: () => {
-          this.windowService.openExternalUrl(`https://github.com/noovolari/leapp/issues/new?labels=bug&body=${encodeURIComponent(this.issueBody)}`);
+          this.windowService.openExternalUrl(
+            `https://github.com/noovolari/leapp/issues/new?labels=bug&body=${encodeURIComponent(this.appService.issueBody)}`
+          );
         },
       },
       { type: "separator" },
@@ -211,76 +207,6 @@ export class TrayMenuComponent implements OnInit, OnDestroy {
 
   ngOnDestroy(): void {
     this.subscribed.unsubscribe();
-  }
-
-  private getMetadata() {
-    const printError = (error) => {
-      this.loggingService.log(new LoggedException(error.toString(), this, LogLevel.error, false, error.stack));
-      if ((error as LeappBaseError).severity === LogLevel.error) {
-        if ((error as LeappLinkError).link === undefined || (error as LeappLinkError).link === null) {
-          this.messageToasterService.toast(error.toString(), ToastLevel.error, "");
-        } else {
-          this.messageToasterService.toast(error.toString(), ToastLevel.error, "", (error as LeappLinkError).link);
-        }
-      }
-    };
-
-    this.getAwsCliVersion()
-      .then(() => {
-        this.getSessionManagerPluginVersion()
-          .then(() => {
-            this.setIssueBody();
-          })
-          .catch((error) => {
-            printError(error);
-          });
-      })
-      .catch((error) => {
-        printError(error);
-      });
-  }
-
-  private async getAwsCliVersion(): Promise<void> {
-    if (!this.awsCliVersion) {
-      try {
-        this.awsCliVersion = await this.appProviderService.executeService.execute("aws --version");
-      } catch (_) {
-        throw new LeappLinkError(
-          "https://docs.leapp.cloud/latest/troubleshooting/faq/",
-          this,
-          "An error occurred getting AWS CLI version. Please check if it is installed and <span class='link'>add a symlink</span> following the documentation."
-        );
-      }
-    }
-  }
-
-  private async getSessionManagerPluginVersion(): Promise<void> {
-    if (!this.awsSsmPluginVersion) {
-      try {
-        const sessionManagerPluginVersion = await this.appProviderService.executeService.execute("session-manager-plugin --version");
-        this.awsSsmPluginVersion = sessionManagerPluginVersion.replace(/(\r\n|\n|\r)/gm, "");
-      } catch (error) {
-        throw new LeappLinkError(
-          "https://docs.leapp.cloud/latest/built-in-features/aws-ec2-connect/",
-          this,
-          "An error occurred getting AWS Session Manager Plugin version. <span class='link'>Click here to follow the instructions on the docs</span> and solve the issue."
-        );
-      }
-    }
-  }
-
-  private setIssueBody(): void {
-    this.issueBody = `### Description:
-> Please include a detailed description of the issue (and an image or screen recording, if applicable)
-
-
-### Details:
-| Leapp Version | ${this.electronService.app.getVersion()} |
-| --- | --- |
-| SsmPluginVersion | ${this.awsSsmPluginVersion} |
-| Platform | ${process.platform} |
-| Awscli | ${this.awsCliVersion}
-`;
   }
 
   private createSessionVoice(session: Session) {